+++ slack/bot.go
@@ -128,46 +128,3 @@ func (r *rtmBot) handleMessage(ev *slack.MessageEvent) error {
 	return nil
 
 }
-
-func (r *rtmBot) handleHelp(channelID string) error {
-	_, _, err := r.rtm.PostMessage(
-		channelID,
-		"Use `@"+r.rtm.GetInfo().User.Name+" nominate` to nominate someone",
-		slack.PostMessageParameters{
-			LinkNames: 1,
-			Markdown:  true,
-		},
-	)
-
-	return err
-}
-
-func (r *rtmBot) handleNominate(callerID, channelID string) error {
-	userIDs, _, err := r.rtm.GetUsersInConversation(
-		&slack.GetUsersInConversationParameters{ChannelID: channelID},
-	)
-	if err != nil {
-		return err
-	}
-
-	botID := r.rtm.GetInfo().User.ID
-	userIDs = filter(userIDs, botID, callerID)
-
-	if len(userIDs) == 0 {
-		_, _, err = r.rtm.PostMessage(channelID, "Nobody to nominate ¯\\_(ツ)_/¯", slack.PostMessageParameters{})
-
-		return err
-	}
-
-	userID := userIDs[r.gen.Intn(len(userIDs))]
-
-	user, err := r.rtm.GetUserInfo(userID)
-
-	if err != nil {
-		return err
-	}
-
-	_, _, err = r.rtm.PostMessage(channelID, "@"+user.Name, slack.PostMessageParameters{LinkNames: 1})
-
-	return err
-}
